Moved Roles tests into new structure
