+++ tests/protocols/gmpv208/entities/roles/test_get_role.py
@@ -19,17 +19,19 @@
 from gvm.errors import RequiredArgument
 
 
-class GmpCloneRoleTestCase:
-    def test_clone(self):
-        self.gmp.clone_role('a1')
+class GmpGetRoleTestMixin:
+    def test_get_role(self):
+        self.gmp.get_role('r1')
 
-        self.connection.send.has_been_called_with(
-            '<create_role>' '<copy>a1</copy>' '</create_role>'
-        )
+        self.connection.send.has_been_called_with('<get_roles role_id="r1"/>')
 
-    def test_missing_id(self):
+        self.gmp.get_role(role_id='r1')
+
+        self.connection.send.has_been_called_with('<get_roles role_id="r1"/>')
+
+    def test_get_role_missing_role_id(self):
         with self.assertRaises(RequiredArgument):
-            self.gmp.clone_role('')
+            self.gmp.get_role(role_id=None)
 
         with self.assertRaises(RequiredArgument):
-            self.gmp.clone_role(None)
+            self.gmp.get_role('')
