+++ tests/protocols/gmpv208/testcmds/__init__.py
@@ -29,19 +29,14 @@ from .test_get_vulnerability import GmpGetVulnerabilityTestCase
 from .test_protocol_version import GmpProtocolVersionTestCase
 from .test_authenticate import GmpAuthenticateTestCase
 from .test_clone_group import GmpCloneGroupTestCase
-from .test_clone_role import GmpCloneRoleTestCase
 from .test_create_group import GmpCreateGroupTestCase
-from .test_create_role import GmpCreateRoleTestCase
 from .test_delete_group import GmpDeleteGroupTestCase
-from .test_delete_role import GmpDeleteRoleTestCase
 from .test_describe_auth import GmpDescribeAuthCommandTestCase
 from .test_empty_trashcan import GmpEmptyTrashcanCommandTestCase
 from .test_get_group import GmpGetGroupTestCase
 from .test_get_groups import GmpGetGroupsTestCase
 from .test_get_preference import GmpGetPreferenceTestCase
 from .test_get_preferences import GmpGetPreferencesTestCase
-from .test_get_role import GmpGetRoleTestCase
-from .test_get_roles import GmpGetRolesTestCase
 from .test_get_setting import GmpGetSettingTestCase
 from .test_get_settings import GmpGetSettingsTestCase
 from .test_get_system_reports import GmpGetSystemReportsTestCase
@@ -49,7 +44,6 @@ from .test_get_version import GmpGetVersionCommandTestCase
 from .test_help import GmpHelpTestCase
 from .test_modify_auth import GmpModifyAuthTestCase
 from .test_modify_group import GmpModifyGroupTestCase
-from .test_modify_role import GmpModifyRoleTestCase
 from .test_modify_setting import GmpModifySettingTestCase
 from .test_restore import GmpRestoreTestCase
 from .test_sync_cert import GmpSyncCertCommandTestCase
